ml-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
hmmlearn==0.3.0
ta==0.11.0
newsapi-python==0.2.7
# CPU-only torch (smaller, works on Railway free tier)
torch==2.2.0+cpu
transformers==4.37.2
optuna==3.5.0
stable-baselines3==2.2.1
gymnasium==0.29.1
joblib==1.3.2
pymongo==4.6.1
motor==3.3.2
httpx==0.26.0
pydantic==2.5.3
python-multipart==0.0.9
requests==2.31.0
feedparser==6.0.11
pyotp==2.9.0
